Renamed currentState to nextState and previousState to currentState
The old model of states was that we were updating the current state and that all finalized state were considered previous.
In the new model of state we consider the updating state to be the "next" state and the current state to be the most recently finished state.

Update NomadsGameVisuals.java